# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WeDo Backend API",
    "version": "1.0.0",
    "description": "OpenAPI specification for the WeDo backend."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["code", "message"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorDetail" }
        },
        "required": ["error"]
      },
      "City": {
        "type": "object",
        "properties": {
          "settlement": { "type": "string", "nullable": true },
          "region": { "type": "string", "nullable": true },
          "country": { "type": "string", "nullable": true },
          "latitude": { "type": "number", "nullable": true },
          "longitude": { "type": "number", "nullable": true },
          "title": { "type": "string", "nullable": true }
        }
      },
      "Privacy": {
        "type": "object",
        "properties": {
          "showAvatar": { "type": "boolean" },
          "showGender": { "type": "boolean" },
          "showCity": { "type": "boolean" },
          "showInterests": { "type": "boolean" },
          "showBirthDate": { "type": "boolean" },
          "showAttendanceHistory": { "type": "boolean" },
          "showReviews": { "type": "boolean" }
        }
      },
      "UserSnippet": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "avatarFileId": { "type": "string", "nullable": true },
          "rating": { "type": "number" }
        },
        "required": ["id", "name", "rating"]
      },
      "AttendanceHistory": {
        "type": "object",
        "properties": {
          "attended": { "type": "integer" },
          "missed": { "type": "integer" }
        }
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "avatarFileId": { "type": "string", "nullable": true },
          "rating": { "type": "number" },
          "age": { "type": "integer", "nullable": true },
          "gender": { "type": "string", "nullable": true },
          "city": { "$ref": "#/components/schemas/City" },
          "interests": { "type": "array", "items": { "type": "string" }, "nullable": true },
          "attendanceHistory": { "allOf": [{ "$ref": "#/components/schemas/AttendanceHistory" }], "nullable": true },
          "reviewsPreview": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "fromUserId": { "type": "string" },
                "fromUserName": { "type": "string" },
                "rating": { "type": "number" },
                "text": { "type": "string" },
                "date": { "type": "string", "format": "date-time" },
                "activityId": { "type": "string" }
              },
              "required": ["id", "fromUserId", "fromUserName", "rating", "text", "date", "activityId"]
            }
          },
          "privacy": { "allOf": [{ "$ref": "#/components/schemas/Privacy" }], "nullable": true },
          "isCurrentUser": { "type": "boolean" },
          "isSubscribed": { "type": "boolean", "nullable": true }
        },
        "required": ["id", "name", "rating", "isCurrentUser"]
      },
      "AuthTokens": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "expiresAt": { "type": "string", "format": "date-time" }
        },
        "required": ["accessToken", "refreshToken", "expiresAt"]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/UserProfile" },
          "tokens": { "$ref": "#/components/schemas/AuthTokens" }
        },
        "required": ["user", "tokens"]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "birthDate": { "type": "string", "format": "date" },
          "gender": { "type": "string", "enum": ["male", "female", "notgiven"] },
          "city": { "$ref": "#/components/schemas/City" },
          "interests": { "type": "array", "items": { "type": "string" } },
          "privacy": { "allOf": [{ "$ref": "#/components/schemas/Privacy" }], "nullable": true }
        },
        "required": ["name", "email", "password", "birthDate", "gender", "city"]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        },
        "required": ["email", "password"]
      },
      "UpdateMeRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "avatarFileId": { "type": "integer", "nullable": true },
          "birthDate": { "type": "string", "format": "date" },
          "gender": { "type": "string", "enum": ["male", "female", "notgiven"] },
          "city": { "$ref": "#/components/schemas/City" },
          "interests": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "QrTokenResolveRequest": {
        "type": "object",
        "properties": {
          "token": { "type": "string" }
        },
        "required": ["token"]
      },
...
      "QrTokenResolveResponse": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/UserSnippet" },
          "expiresAt": { "type": "string", "format": "date-time" }
        },
        "required": ["user", "expiresAt"]
      },
      "ActivityLocation": {
        "type": "object",
        "properties": {
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "address": { "type": "string" },
          "name": { "type": "string", "nullable": true },
          "settlement": { "type": "string", "nullable": true },
          "region": { "type": "string", "nullable": true },
          "country": { "type": "string", "nullable": true }
        },
        "required": ["latitude", "longitude", "address"]
      },
      "ActivityPreferences": {
        "type": "object",
        "properties": {
          "gender": { "type": "string", "enum": ["male", "female"], "nullable": true },
          "ageFrom": { "type": "integer", "nullable": true },
          "ageTo": { "type": "integer", "nullable": true },
          "level": { "type": "string", "enum": ["beginner", "intermediate", "advanced"], "nullable": true },
          "maxParticipants": { "type": "integer", "nullable": true }
        }
      },
      "ActivityListItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "startAt": { "type": "string", "format": "date-time" },
          "endAt": { "type": "string", "format": "date-time" },
          "format": { "type": "string", "enum": ["online", "offline"] },
          "status": { "type": "string", "enum": ["active", "cancelled"] },
          "location": { "$ref": "#/components/schemas/ActivityLocation" },
          "categoryId": { "type": "string" },
          "subcategoryId": { "type": "string", "nullable": true },
          "coverPhotoFileId": { "type": "string", "nullable": true },
          "photoFileIds": { "type": "array", "items": { "type": "string" } },
          "organizer": { "$ref": "#/components/schemas/UserSnippet" },
          "participantsCount": { "type": "integer" },
          "maxParticipants": { "type": "integer", "nullable": true },
          "requiresApproval": { "type": "boolean" },
          "preferences": { "allOf": [{ "$ref": "#/components/schemas/ActivityPreferences" }], "nullable": true },
          "pendingRequestsCount": { "type": "integer", "nullable": true },
          "price": { "type": "number" }
        },
        "required": ["id", "title", "startAt", "endAt", "format", "status", "location", "categoryId", "photoFileIds", "organizer", "participantsCount", "requiresApproval", "price"]
      },
...
      "ActivityListResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/ActivityListItem" } },
          "nextCursor": { "type": "string", "nullable": true },
          "hasMore": { "type": "boolean" }
        },
        "required": ["items", "hasMore"]
      },
...
    }
  },
  "security": [
    { "BearerAuth": [] }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register a new user",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login with email and password",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["auth"],
        "summary": "Logout current user",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Logged out successfully" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "summary": "Refresh access token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": { "type": "string" }
                },
                "required": ["refreshToken"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New tokens",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthTokens" }
              }
            }
          },
          "401": {
            "description": "Invalid token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "tags": ["users"],
        "summary": "Get current user profile",
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserProfile" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["users"],
        "summary": "Update current user profile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateMeRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserProfile" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Delete current user account",
        "responses": {
          "204": { "description": "Account deleted" }
        }
      }
    },
...
    "/me/my-activities": {
      "get": {
        "tags": ["users"],
        "summary": "Get my activities",
        "parameters": [
          { "name": "tab", "in": "query", "schema": { "type": "string", "enum": ["created", "upcoming", "attended"] } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "My activities page",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ActivityListResponse" }
              }
            }
          }
        }
      }
    },
...
    "/users/{user_id}": {
      "get": {
        "tags": ["users"],
        "summary": "Get user profile by id",
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserProfile" }
              }
            }
          }
        }
      }
    },
...
    "/users/{user_id}/rating": {
      "get": {
        "tags": ["users"],
        "summary": "Get user rating",
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "User rating",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rating": { "type": "number" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/attendance-history": {
      "get": {
        "tags": ["users"],
        "summary": "Get user attendance history",
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Attendance history",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AttendanceHistory" }
              }
            }
          }
        }
      }
    },
...
    "/qr-tokens/resolve": {
      "post": {
        "tags": ["users"],
        "summary": "Resolve QR token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QrTokenResolveRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resolved token data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/QrTokenResolveResponse" }
              }
            }
          }
        }
      }
    },
...
    "/activities/{activity_id}/attendance/scan": {
      "post": {
        "tags": ["participation"],
        "summary": "Mark attendance via QR token",
        "parameters": [
          { "name": "activity_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QrTokenResolveRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "Success" }
        }
      }
    },
...
  }
}